documentation

+++ process_multiparametric.py
@@ -8,9 +8,11 @@ intermediate_filename = 'multiparam.tmp'
 run_args = ['python3', 'create_universal_HDF5.py', '-i', intermediate_filename]
 
 def help():
-    print("usage: python create_universal_HDF5.py -i-reg [Free-form input] -i-mp [multiparam FORTRAN input] -univ-inps [usual inputs]\n"
+    print("usage: python create_universal_HDF5.py -i-reg [Free-form input] -i-mp [multiparam FORTRAN input] [-keep-intermediate]\n"
+          "                                       -univ-inps [usual inputs]\n"
           "       The inputs are the free-form input file and the driving file for FORTRAN, all the inputs of the universal input \n"
           '       (except the input file) must be placed as last arguments after the flag "-univ-inps".')
+          '       ("-keep-intermediate" is a debugging flag to keep intermediate files.)')
     exit(0)
 
 ## Treat arguments
@@ -23,6 +25,8 @@ else:
     inputfilenamemp = arguments[arg_index+1]
     arg_index = arguments.index("-univ-inps")
     run_args = run_args + arguments[(arg_index+1):]
+    keep_intermadiate = ("-keep-intermediate" in arguments)
+
 
 ## Prepare files for the FORTRAN code
 with open(inputfilenamemp, "r") as InputMP:
@@ -52,12 +56,12 @@ with open('multiparam_FORTRAN.inp','w') as f:
 
 ## run FORTRAN
 subprocess.run('./all_combinations.e')
-os.remove('multiparam_FORTRAN.inp')
+if (not keep_intermadiate): os.remove('multiparam_FORTRAN.inp')
 
 ## append FORTRAN output to the universal-input driving file
 with open('list_of_combinations_FORTRAN.dat') as f:
     content = f.read()
-os.remove('list_of_combinations_FORTRAN.dat')
+if (not keep_intermadiate): os.remove('list_of_combinations_FORTRAN.dat')
 
 content = '$multiparametric\t' + str(content.count('\n')) + '\n' + names + '\n' + dtypes + '\n' + units + '\n' + content
 
@@ -66,5 +70,5 @@ with open(intermediate_filename,'w') as f_tmp, open(inputfilename,'r') as f_reg:
 
 ## run the universal-input
 subprocess.run(run_args)
-os.remove(intermediate_filename)
+if (not keep_intermadiate): os.remove(intermediate_filename)
 # https://stackoverflow.com/questions/7152340/using-a-python-subprocess-call-to-invoke-a-python-script
